refactor(pine): move the built-in families out of the evaluator core
`eval.rs` was 1,669 production lines: the tree walk and every built-in
family it dispatches to lived in one file, so an agent asked to change
`ta.sma`'s ring read the loop budget, the draw-object cap and the switch
arms on the way.

Three siblings now own one family each, by pure move — `eval/kernels.rs`
the stateful `ta.*` kernels, `eval/pure.rs` the pure math, colour and
string built-ins, and `eval/objects.rs` `line.new`/`box.new`/`label.new`
with the argument coercions a drawing takes and the `handle.set_*`
methods, plus the four `DEFAULT_*` constants only it reads. `eval.rs`
keeps the vocabulary and the walk: `Value`, `ScriptState`, `Kernel`,
`Eval` and the statement/expression evaluator that routes a call.

Every body moved unchanged. The only widenings are `pub(super)` on the
six methods `builtin_call` and `eval` still call across the seam.
`eval.rs` falls to 974 production lines; the largest sibling is 343.
