+++ crates/pine/src/eval.rs
@@ -20,8 +20,7 @@ use std::collections::BTreeMap;
 use std::rc::Rc;
 
 use quantick_indicators::{
-    BoxObj, Ctx, IndicatorBar, InputValue, LabelObj, LabelStyle, LineObj, MarkerLocation, ObjectId,
-    ObjectStore, Rgba8, SourceId, fmath, ta,
+    Ctx, IndicatorBar, InputValue, MarkerLocation, ObjectId, ObjectStore, Rgba8, SourceId, ta,
 };
 
 use crate::ast::{BinOp, NodeId, NodeKind, UnOp, VarMode};
@@ -29,18 +28,13 @@ use crate::builtins::Builtin;
 use crate::compile::{CompiledScript, Resolution};
 use crate::error::{ErrorCode, PineError, Span};
 
+mod kernels;
+mod objects;
+mod pure;
+
 /// Iteration budget per `for`/`while` per bar — a runaway loop errors the
 /// indicator instead of freezing the worker (§3.5).
 pub const LOOP_BUDGET: u32 = 10_000;
-
-/// Default draw-object stroke/border/label color (a readable blue).
-const DEFAULT_OBJECT_COLOR: Rgba8 = Rgba8::opaque(41, 98, 255);
-/// Default box fill: the border color at ~85% transparency.
-const DEFAULT_BOX_FILL: Rgba8 = Rgba8::new(41, 98, 255, 38);
-/// Default label text color.
-const DEFAULT_LABEL_TEXT: Rgba8 = Rgba8::opaque(255, 255, 255);
-/// Default line width in points.
-const DEFAULT_LINE_WIDTH: f32 = 1.0;
 
 /// A runtime value. `na` for numbers is NaN inside `Num`; the distinct `Na`
 /// covers non-numeric contexts.
@@ -930,695 +924,6 @@ impl<'a> Eval<'a> {
             Some(arg) => self.num(arg.value),
             None => Ok(f64::NAN),
         }
-    }
-
-    fn kernel_call(
-        &mut self,
-        id: NodeId,
-        builtin: Builtin,
-        site: u32,
-        args: &[crate::ast::Arg],
-    ) -> Result<Value, PineError> {
-        let key = (site, self.path);
-        // Bool-family kernels take their own paths below.
-        match builtin {
-            Builtin::TaCrossover | Builtin::TaCrossunder | Builtin::TaCross => {
-                let a = self.arg_num(args, 0)?;
-                let b = self.arg_num(args, 1)?;
-                let kernel = self
-                    .state
-                    .kernels
-                    .entry(key)
-                    .or_insert_with(|| match builtin {
-                        Builtin::TaCrossover => Kernel::CrossOver(ta::CrossOver::new()),
-                        Builtin::TaCrossunder => Kernel::CrossUnder(ta::CrossUnder::new()),
-                        _ => Kernel::Cross(ta::Cross::new()),
-                    });
-                let fired = match kernel {
-                    Kernel::CrossOver(k) => k.push(a, b),
-                    Kernel::CrossUnder(k) => k.push(a, b),
-                    Kernel::Cross(k) => k.push(a, b),
-                    _ => unreachable!("cross site holds a cross kernel"),
-                };
-                return Ok(Value::Bool(fired));
-            }
-            Builtin::TaBarsSince => {
-                let condition = match args.first() {
-                    Some(arg) => self.cond(arg.value)?,
-                    None => false,
-                };
-                let kernel = self
-                    .state
-                    .kernels
-                    .entry(key)
-                    .or_insert_with(|| Kernel::BarsSince(ta::BarsSince::new()));
-                let Kernel::BarsSince(k) = kernel else {
-                    unreachable!("barssince site holds a barssince kernel");
-                };
-                return Ok(Value::Num(k.push(condition)));
-            }
-            Builtin::TaValueWhen => {
-                let condition = match args.first() {
-                    Some(arg) => self.cond(arg.value)?,
-                    None => false,
-                };
-                let source = self.arg_num(args, 1)?;
-                let occurrence = self.arg_num(args, 2)?;
-                #[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
-                let occurrence = if occurrence.is_nan() {
-                    0
-                } else {
-                    occurrence.max(0.0) as usize
-                };
-                let kernel = self
-                    .state
-                    .kernels
-                    .entry(key)
-                    .or_insert_with(|| Kernel::ValueWhen(ta::ValueWhen::new(occurrence)));
-                let Kernel::ValueWhen(k) = kernel else {
-                    unreachable!("valuewhen site holds a valuewhen kernel");
-                };
-                return Ok(Value::Num(k.push(condition, source)));
-            }
-            Builtin::Fixnan => {
-                let x = self.arg_num(args, 0)?;
-                let kernel = self
-                    .state
-                    .kernels
-                    .entry(key)
-                    .or_insert_with(|| Kernel::Fixnan(f64::NAN));
-                let Kernel::Fixnan(last) = kernel else {
-                    unreachable!("fixnan site holds a fixnan kernel");
-                };
-                if !x.is_nan() {
-                    *last = x;
-                }
-                return Ok(Value::Num(*last));
-            }
-            _ => {}
-        }
-
-        // Numeric kernels: gather inputs, then a positive-int length where
-        // the kernel takes one (re-validated here because inputs bind per
-        // instance).
-        let length = |v: f64, id: NodeId, this: &Self| -> Result<usize, PineError> {
-            // The upper bound matters as much as the lower one: a kernel
-            // allocates its ring up front, and a failed allocation aborts the
-            // process instead of unwinding into this indicator's error state.
-            // Inputs are bound per instance, so the load-time check in
-            // `compile` is not the last word — this is.
-            let in_range = v.is_finite()
-                && v >= 1.0
-                && v.fract() == 0.0
-                && v <= crate::compile::MAX_KERNEL_LENGTH as f64;
-            if in_range {
-                #[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
-                Ok(v as usize)
-            } else {
-                Err(this.err(
-                    id,
-                    ErrorCode::PineSeriesLength,
-                    format!(
-                        "kernel length must be a positive integer no greater than {}, got {v}",
-                        crate::compile::MAX_KERNEL_LENGTH
-                    ),
-                ))
-            }
-        };
-        let value = match builtin {
-            Builtin::TaTr => {
-                let handle_na = match args.first() {
-                    Some(arg) => self.cond(arg.value)?,
-                    None => false,
-                };
-                let (high, low, close) = (self.bar.high, self.bar.low, self.bar.close);
-                let kernel = self
-                    .state
-                    .kernels
-                    .entry(key)
-                    .or_insert_with(|| Kernel::Tr(ta::Tr::new(handle_na)));
-                let Kernel::Tr(k) = kernel else {
-                    unreachable!("tr site holds a tr kernel");
-                };
-                k.push(high, low, close)
-            }
-            Builtin::TaAtr => {
-                let len = length(self.arg_num(args, 0)?, id, self)?;
-                self.pin_kernel_length(key, len, id)?;
-                let (high, low, close) = (self.bar.high, self.bar.low, self.bar.close);
-                let kernel = self
-                    .state
-                    .kernels
-                    .entry(key)
-                    .or_insert_with(|| Kernel::Atr(ta::Atr::new(len)));
-                let Kernel::Atr(k) = kernel else {
-                    unreachable!("atr site holds an atr kernel");
-                };
-                k.push(high, low, close)
-            }
-            Builtin::TaCum => {
-                let x = self.arg_num(args, 0)?;
-                let kernel = self
-                    .state
-                    .kernels
-                    .entry(key)
-                    .or_insert_with(|| Kernel::Cum(ta::Cum::new()));
-                let Kernel::Cum(k) = kernel else {
-                    unreachable!("cum site holds a cum kernel");
-                };
-                k.push(x)
-            }
-            Builtin::TaChange | Builtin::TaMom => {
-                let x = self.arg_num(args, 0)?;
-                let n = match args.get(1) {
-                    Some(arg) => length(self.num(arg.value)?, id, self)?,
-                    None => 1,
-                };
-                self.pin_kernel_length(key, n, id)?;
-                let kernel = self
-                    .state
-                    .kernels
-                    .entry(key)
-                    .or_insert_with(|| Kernel::Change(ta::Change::new(n)));
-                let Kernel::Change(k) = kernel else {
-                    unreachable!("change site holds a change kernel");
-                };
-                k.push(x)
-            }
-            Builtin::TaVwma => {
-                let src = self.arg_num(args, 0)?;
-                let len = length(self.arg_num(args, 1)?, id, self)?;
-                self.pin_kernel_length(key, len, id)?;
-                let volume = self.bar.volume();
-                let kernel = self
-                    .state
-                    .kernels
-                    .entry(key)
-                    .or_insert_with(|| Kernel::Vwma(ta::Vwma::new(len)));
-                let Kernel::Vwma(k) = kernel else {
-                    unreachable!("vwma site holds a vwma kernel");
-                };
-                k.push(src, volume)
-            }
-            Builtin::TaPivotHigh | Builtin::TaPivotLow => {
-                let src = self.arg_num(args, 0)?;
-                let left = length(self.arg_num(args, 1)?, id, self)?;
-                let right = length(self.arg_num(args, 2)?, id, self)?;
-                self.pin_kernel_length(key, left + right, id)?;
-                let is_high = builtin == Builtin::TaPivotHigh;
-                let kernel = self.state.kernels.entry(key).or_insert_with(|| {
-                    if is_high {
-                        Kernel::PivotHigh(ta::PivotHigh::new(left, right))
-                    } else {
-                        Kernel::PivotLow(ta::PivotLow::new(left, right))
-                    }
-                });
-                match kernel {
-                    Kernel::PivotHigh(k) => k.push(src),
-                    Kernel::PivotLow(k) => k.push(src),
-                    _ => unreachable!("pivot site holds a pivot kernel"),
-                }
-            }
-            // The (src, len) family.
-            _ => {
-                let src = self.arg_num(args, 0)?;
-                let len = length(self.arg_num(args, 1)?, id, self)?;
-                self.pin_kernel_length(key, len, id)?;
-                let kernel = self
-                    .state
-                    .kernels
-                    .entry(key)
-                    .or_insert_with(|| match builtin {
-                        Builtin::TaSma => Kernel::Sma(ta::Sma::new(len)),
-                        Builtin::TaEma => Kernel::Ema(ta::Ema::new(len)),
-                        Builtin::TaRma => Kernel::Rma(ta::Rma::new(len)),
-                        Builtin::TaWma => Kernel::Wma(ta::Wma::new(len)),
-                        Builtin::TaRsi => Kernel::Rsi(ta::Rsi::new(len)),
-                        Builtin::TaStdev => Kernel::Stdev(ta::Stdev::new(len)),
-                        Builtin::TaHighest => Kernel::Highest(ta::Highest::new(len)),
-                        Builtin::TaLowest => Kernel::Lowest(ta::Lowest::new(len)),
-                        Builtin::TaHighestBars => Kernel::HighestBars(ta::HighestBars::new(len)),
-                        Builtin::TaLowestBars => Kernel::LowestBars(ta::LowestBars::new(len)),
-                        Builtin::MathSum => Kernel::WindowSum(ta::WindowSum::new(len)),
-                        other => unreachable!("{other:?} is not a (src, len) kernel"),
-                    });
-                match kernel {
-                    Kernel::Sma(k) => k.push(src),
-                    Kernel::Ema(k) => k.push(src),
-                    Kernel::Rma(k) => k.push(src),
-                    Kernel::Wma(k) => k.push(src),
-                    Kernel::Rsi(k) => k.push(src),
-                    Kernel::Stdev(k) => k.push(src),
-                    Kernel::Highest(k) => k.push(src),
-                    Kernel::Lowest(k) => k.push(src),
-                    Kernel::HighestBars(k) => k.push(src),
-                    Kernel::LowestBars(k) => k.push(src),
-                    Kernel::WindowSum(k) => k.push(src),
-                    _ => unreachable!("(src, len) site holds a (src, len) kernel"),
-                }
-            }
-        };
-        Ok(Value::Num(value))
-    }
-
-    fn pure_call(
-        &mut self,
-        id: NodeId,
-        builtin: Builtin,
-        args: &[crate::ast::Arg],
-    ) -> Result<Value, PineError> {
-        let value = match builtin {
-            Builtin::NaCall => {
-                let v = match args.first() {
-                    Some(arg) => self.eval(arg.value)?,
-                    None => Value::Na,
-                };
-                return Ok(Value::Bool(v.is_na()));
-            }
-            Builtin::Nz => {
-                let v = match args.first() {
-                    Some(arg) => self.eval(arg.value)?,
-                    None => Value::Na,
-                };
-                if v.is_na() {
-                    return match args.get(1) {
-                        Some(replacement) => self.eval(replacement.value),
-                        None => Ok(Value::Num(0.0)),
-                    };
-                }
-                return Ok(v);
-            }
-            Builtin::ColorNew => {
-                let base = match args.first() {
-                    Some(arg) => self.eval(arg.value)?,
-                    None => Value::Na,
-                };
-                let Value::Color(rgba) = base else {
-                    return Err(self.type_err(id, "a color", &base));
-                };
-                let transparency = self.arg_num(args, 1)?;
-                #[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
-                let alpha = (255.0 * (1.0 - (transparency / 100.0).clamp(0.0, 1.0))).round() as u32;
-                return Ok(Value::Color((rgba & 0xFFFF_FF00) | alpha));
-            }
-            Builtin::ColorRgb => {
-                let r = self.arg_num(args, 0)?;
-                let g = self.arg_num(args, 1)?;
-                let b = self.arg_num(args, 2)?;
-                let transparency = match args.get(3) {
-                    Some(arg) => self.num(arg.value)?,
-                    None => 0.0,
-                };
-                #[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
-                let pack = |v: f64| (v.clamp(0.0, 255.0)) as u32;
-                #[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
-                let alpha = (255.0 * (1.0 - (transparency / 100.0).clamp(0.0, 1.0))).round() as u32;
-                return Ok(Value::Color(
-                    (pack(r) << 24) | (pack(g) << 16) | (pack(b) << 8) | alpha,
-                ));
-            }
-            Builtin::MathAbs => self.arg_num(args, 0)?.abs(),
-            Builtin::MathFloor => self.arg_num(args, 0)?.floor(),
-            Builtin::MathCeil => self.arg_num(args, 0)?.ceil(),
-            Builtin::MathRound => self.arg_num(args, 0)?.round(),
-            Builtin::MathSign => {
-                let v = self.arg_num(args, 0)?;
-                if v.is_nan() {
-                    f64::NAN
-                } else if v > 0.0 {
-                    1.0
-                } else if v < 0.0 {
-                    -1.0
-                } else {
-                    0.0
-                }
-            }
-            Builtin::MathSqrt => self.arg_num(args, 0)?.sqrt(),
-            Builtin::MathMax | Builtin::MathMin | Builtin::MathAvg => {
-                // Streaming, so the per-bar path allocates nothing — these
-                // appear inside plotted expressions. The identities are the
-                // infinities, not `f64::MIN`/`MAX`: those are the extreme
-                // *finite* values, so `math.max(math.log(0))` answered
-                // -1.797e308 where -inf is the true maximum of its arguments.
-                let mut max = f64::NEG_INFINITY;
-                let mut min = f64::INFINITY;
-                let mut sum = 0.0;
-                let mut count = 0usize;
-                let mut saw_nan = false;
-                for arg in args {
-                    let v = self.num(arg.value)?;
-                    if v.is_nan() {
-                        saw_nan = true;
-                    }
-                    max = max.max(v);
-                    min = min.min(v);
-                    sum += v;
-                    count += 1;
-                }
-                if count == 0 || saw_nan {
-                    f64::NAN
-                } else {
-                    match builtin {
-                        Builtin::MathMax => max,
-                        Builtin::MathMin => min,
-                        _ => sum / count as f64,
-                    }
-                }
-            }
-            Builtin::MathPow => fmath::pow(self.arg_num(args, 0)?, self.arg_num(args, 1)?),
-            Builtin::MathExp => fmath::exp(self.arg_num(args, 0)?),
-            Builtin::MathLog => fmath::ln(self.arg_num(args, 0)?),
-            Builtin::MathLog10 => fmath::log10(self.arg_num(args, 0)?),
-            other => {
-                return Err(self.err(
-                    id,
-                    ErrorCode::PineType,
-                    format!("`{other:?}` cannot be called here"),
-                ));
-            }
-        };
-        Ok(Value::Num(value))
-    }
-
-    // ---- draw objects ------------------------------------------------
-
-    /// A named-or-positional argument, mirroring the compile-time helper.
-    fn call_arg<'b>(
-        args: &'b [crate::ast::Arg],
-        position: usize,
-        name: &str,
-    ) -> Option<&'b crate::ast::Arg> {
-        args.iter()
-            .find(|a| a.name.as_deref() == Some(name))
-            .or_else(|| args.iter().filter(|a| a.name.is_none()).nth(position))
-    }
-
-    fn arg_bar_index(
-        &mut self,
-        args: &[crate::ast::Arg],
-        position: usize,
-        name: &str,
-    ) -> Result<i64, PineError> {
-        match Self::call_arg(args, position, name) {
-            Some(arg) => {
-                let v = self.num(arg.value)?;
-                #[allow(clippy::cast_possible_truncation)]
-                Ok(if v.is_finite() {
-                    v.round() as i64
-                } else {
-                    // `label.new(na, price)` has no bar to sit on. Zero would
-                    // anchor it to the very first bar — a real label drawn
-                    // out of a missing value, while the price side of the
-                    // same call answers NaN and the renderer honestly skips
-                    // it. `OFF_CHART_BAR` is negative, which the renderer's
-                    // existing `x < 0` guards already drop.
-                    quantick_indicators::OFF_CHART_BAR
-                })
-            }
-            None => Ok(self.ctx.bar_index as i64),
-        }
-    }
-
-    fn arg_price(
-        &mut self,
-        args: &[crate::ast::Arg],
-        position: usize,
-        name: &str,
-    ) -> Result<f64, PineError> {
-        match Self::call_arg(args, position, name) {
-            Some(arg) => self.num(arg.value),
-            None => Ok(f64::NAN),
-        }
-    }
-
-    fn arg_color(
-        &mut self,
-        args: &[crate::ast::Arg],
-        position: usize,
-        name: &str,
-        default: Rgba8,
-    ) -> Result<Rgba8, PineError> {
-        match Self::call_arg(args, position, name) {
-            Some(arg) => {
-                let v = self.eval(arg.value)?;
-                match v {
-                    Value::Color(rgba) => Ok(Rgba8::from_u32(rgba)),
-                    Value::Na => Ok(default),
-                    other => Err(self.type_err(arg.value, "a color", &other)),
-                }
-            }
-            None => Ok(default),
-        }
-    }
-
-    fn arg_text(
-        &mut self,
-        args: &[crate::ast::Arg],
-        position: usize,
-        name: &str,
-    ) -> Result<String, PineError> {
-        match Self::call_arg(args, position, name) {
-            Some(arg) => {
-                let v = self.eval(arg.value)?;
-                match v {
-                    Value::Str(s) => Ok(s.as_ref().clone()),
-                    Value::Na => Ok(String::new()),
-                    other => Err(self.type_err(arg.value, "a string", &other)),
-                }
-            }
-            None => Ok(String::new()),
-        }
-    }
-
-    fn line_new(&mut self, args: &[crate::ast::Arg]) -> Result<Value, PineError> {
-        let x1 = self.arg_bar_index(args, 0, "x1")?;
-        let y1 = self.arg_price(args, 1, "y1")?;
-        let x2 = self.arg_bar_index(args, 2, "x2")?;
-        let y2 = self.arg_price(args, 3, "y2")?;
-        let color = self.arg_color(args, 4, "color", DEFAULT_OBJECT_COLOR)?;
-        let width = match Self::call_arg(args, 5, "width") {
-            Some(arg) => {
-                #[allow(clippy::cast_possible_truncation)]
-                let w = self.num(arg.value)? as f32;
-                w
-            }
-            None => DEFAULT_LINE_WIDTH,
-        };
-        let handle = self.state.objects.new_line(LineObj {
-            x1,
-            y1,
-            x2,
-            y2,
-            color,
-            width,
-        });
-        Ok(Value::Obj(ObjKind::Line, handle.0))
-    }
-
-    fn box_new(&mut self, args: &[crate::ast::Arg]) -> Result<Value, PineError> {
-        let left = self.arg_bar_index(args, 0, "left")?;
-        let top = self.arg_price(args, 1, "top")?;
-        let right = self.arg_bar_index(args, 2, "right")?;
-        let bottom = self.arg_price(args, 3, "bottom")?;
-        let border_color = self.arg_color(args, 4, "border_color", DEFAULT_OBJECT_COLOR)?;
-        let bg_color = self.arg_color(args, 5, "bgcolor", DEFAULT_BOX_FILL)?;
-        let handle = self.state.objects.new_box(BoxObj {
-            left,
-            top,
-            right,
-            bottom,
-            border_color,
-            bg_color,
-            border_width: DEFAULT_LINE_WIDTH,
-        });
-        Ok(Value::Obj(ObjKind::Box, handle.0))
-    }
-
-    fn label_new(&mut self, args: &[crate::ast::Arg]) -> Result<Value, PineError> {
-        let x = self.arg_bar_index(args, 0, "x")?;
-        let y = self.arg_price(args, 1, "y")?;
-        let text = self.arg_text(args, 2, "text")?;
-        let color = self.arg_color(args, 3, "color", DEFAULT_OBJECT_COLOR)?;
-        let text_color = self.arg_color(args, 4, "textcolor", DEFAULT_LABEL_TEXT)?;
-        let style = match Self::call_arg(args, 5, "style") {
-            Some(arg) => {
-                let v = self.eval(arg.value)?;
-                match v {
-                    Value::Str(tag) => match tag.as_str() {
-                        "style_label_up" => LabelStyle::Up,
-                        "style_label_down" => LabelStyle::Down,
-                        _ => LabelStyle::None,
-                    },
-                    _ => LabelStyle::Down,
-                }
-            }
-            None => LabelStyle::Down,
-        };
-        let handle = self.state.objects.new_label(LabelObj {
-            x,
-            y,
-            text,
-            color,
-            text_color,
-            style,
-        });
-        Ok(Value::Obj(ObjKind::Label, handle.0))
-    }
-
-    /// `handle.set_*(...)` / `handle.delete()`. A stale handle (collected
-    /// by the cap, or deleted) is a silent no-op — Pine's behavior, and the
-    /// honest one for GC'd objects.
-    #[allow(clippy::too_many_lines)]
-    fn method_call(
-        &mut self,
-        id: NodeId,
-        kind: ObjKind,
-        handle: ObjectId,
-        field: &str,
-        args: &[crate::ast::Arg],
-    ) -> Result<Value, PineError> {
-        match (kind, field) {
-            (ObjKind::Line, "delete") => self.state.objects.delete_line(handle),
-            (ObjKind::Box, "delete") => self.state.objects.delete_box(handle),
-            (ObjKind::Label, "delete") => self.state.objects.delete_label(handle),
-            (ObjKind::Line, "set_xy1" | "set_xy2") => {
-                let x = self.arg_bar_index(args, 0, "x")?;
-                let y = self.arg_price(args, 1, "y")?;
-                if let Some(line) = self.state.objects.line_mut(handle) {
-                    if field == "set_xy1" {
-                        line.x1 = x;
-                        line.y1 = y;
-                    } else {
-                        line.x2 = x;
-                        line.y2 = y;
-                    }
-                }
-            }
-            (ObjKind::Line, "set_x1" | "set_x2") => {
-                let x = self.arg_bar_index(args, 0, "x")?;
-                if let Some(line) = self.state.objects.line_mut(handle) {
-                    if field == "set_x1" {
-                        line.x1 = x;
-                    } else {
-                        line.x2 = x;
-                    }
-                }
-            }
-            (ObjKind::Line, "set_y1" | "set_y2") => {
-                let y = self.arg_price(args, 0, "y")?;
-                if let Some(line) = self.state.objects.line_mut(handle) {
-                    if field == "set_y1" {
-                        line.y1 = y;
-                    } else {
-                        line.y2 = y;
-                    }
-                }
-            }
-            (ObjKind::Line, "set_color") => {
-                let color = self.arg_color(args, 0, "color", DEFAULT_OBJECT_COLOR)?;
-                if let Some(line) = self.state.objects.line_mut(handle) {
-                    line.color = color;
-                }
-            }
-            (ObjKind::Line, "set_width") => {
-                #[allow(clippy::cast_possible_truncation)]
-                let width = self.arg_price(args, 0, "width")? as f32;
-                if let Some(line) = self.state.objects.line_mut(handle) {
-                    line.width = width;
-                }
-            }
-            (ObjKind::Box, "set_left" | "set_right") => {
-                let x = self.arg_bar_index(args, 0, "x")?;
-                if let Some(object) = self.state.objects.box_mut(handle) {
-                    if field == "set_left" {
-                        object.left = x;
-                    } else {
-                        object.right = x;
-                    }
-                }
-            }
-            (ObjKind::Box, "set_top" | "set_bottom") => {
-                let y = self.arg_price(args, 0, "y")?;
-                if let Some(object) = self.state.objects.box_mut(handle) {
-                    if field == "set_top" {
-                        object.top = y;
-                    } else {
-                        object.bottom = y;
-                    }
-                }
-            }
-            (ObjKind::Box, "set_lefttop" | "set_rightbottom") => {
-                let x = self.arg_bar_index(args, 0, "x")?;
-                let y = self.arg_price(args, 1, "y")?;
-                if let Some(object) = self.state.objects.box_mut(handle) {
-                    if field == "set_lefttop" {
-                        object.left = x;
-                        object.top = y;
-                    } else {
-                        object.right = x;
-                        object.bottom = y;
-                    }
-                }
-            }
-            (ObjKind::Box, "set_bgcolor") => {
-                let color = self.arg_color(args, 0, "color", DEFAULT_BOX_FILL)?;
-                if let Some(object) = self.state.objects.box_mut(handle) {
-                    object.bg_color = color;
-                }
-            }
-            (ObjKind::Box, "set_border_color") => {
-                let color = self.arg_color(args, 0, "color", DEFAULT_OBJECT_COLOR)?;
-                if let Some(object) = self.state.objects.box_mut(handle) {
-                    object.border_color = color;
-                }
-            }
-            (ObjKind::Label, "set_xy") => {
-                let x = self.arg_bar_index(args, 0, "x")?;
-                let y = self.arg_price(args, 1, "y")?;
-                if let Some(label) = self.state.objects.label_mut(handle) {
-                    label.x = x;
-                    label.y = y;
-                }
-            }
-            (ObjKind::Label, "set_x") => {
-                let x = self.arg_bar_index(args, 0, "x")?;
-                if let Some(label) = self.state.objects.label_mut(handle) {
-                    label.x = x;
-                }
-            }
-            (ObjKind::Label, "set_y") => {
-                let y = self.arg_price(args, 0, "y")?;
-                if let Some(label) = self.state.objects.label_mut(handle) {
-                    label.y = y;
-                }
-            }
-            (ObjKind::Label, "set_text") => {
-                let text = self.arg_text(args, 0, "text")?;
-                if let Some(label) = self.state.objects.label_mut(handle) {
-                    label.text = text;
-                }
-            }
-            (ObjKind::Label, "set_color") => {
-                let color = self.arg_color(args, 0, "color", DEFAULT_OBJECT_COLOR)?;
-                if let Some(label) = self.state.objects.label_mut(handle) {
-                    label.color = color;
-                }
-            }
-            (ObjKind::Label, "set_textcolor") => {
-                let color = self.arg_color(args, 0, "textcolor", DEFAULT_LABEL_TEXT)?;
-                if let Some(label) = self.state.objects.label_mut(handle) {
-                    label.text_color = color;
-                }
-            }
-            _ => {
-                return Err(self.err(
-                    id,
-                    ErrorCode::PineType,
-                    format!("unknown method `{field}` for this object kind"),
-                ));
-            }
-        }
-        Ok(Value::Na)
     }
 }
